+++ AI-UI/StableInterface/Structures.cs
@@ -4,7 +4,7 @@ namespace AI_UI {
     public struct RequestStruct {
         public string prompt { get; set; }
         public string negative_prompt { get; set; }
-        public int seed { get; set; }
+        public long seed { get; set; }
         public int steps { get; set; }
         public int width { get; set; }
         public int height { get; set; }
